cifs nfs gmock on bsd

+++ tests/unit/net/tcp_service.cpp
@@ -84,7 +84,7 @@ private:
             ASSERT_GT(writeDescriptorsCount, 0);
             ASSERT_TRUE(FD_ISSET(socket(), &writeDescriptiorsSet));
             ssize_t bytesSent = send(socket(), TestResponse, strlen(TestResponse), MSG_NOSIGNAL);
-            EXPECT_EQ(strlen(TestResponse), bytesSent);
+            EXPECT_EQ(static_cast<ssize_t>(strlen(TestResponse)), bytesSent);
         }
     private:
         TestTcpService& _service;
@@ -111,7 +111,7 @@ TEST(TestTcpService, requestResponse)
     IpEndpoint endpoint{ListenHost, ListenPort};
     ASSERT_EQ(0, connect(s, endpoint.addrinfo()->ai_addr, endpoint.addrinfo()->ai_addrlen));
     ssize_t bytesSent = send(s, TestRequest, strlen(TestRequest), MSG_NOSIGNAL);
-    EXPECT_EQ(strlen(TestRequest), bytesSent);
+    EXPECT_EQ(static_cast<ssize_t>(strlen(TestRequest)), bytesSent);
     char receiveBuffer[ReceiveBufferSize];
     ssize_t bytesReceived = recv(s, receiveBuffer, sizeof(receiveBuffer), 0);
     EXPECT_EQ(TestResponse, std::string(receiveBuffer, bytesReceived));
@@ -138,7 +138,7 @@ TEST(TestTcpService, multipleRequestResponse)
     for (auto & s : sockets)
     {
         ssize_t bytesSent = send(s, TestRequest, strlen(TestRequest), MSG_NOSIGNAL);
-        EXPECT_EQ(strlen(TestRequest), bytesSent);
+        EXPECT_EQ(static_cast<ssize_t>(strlen(TestRequest)), bytesSent);
     }
     char receiveBuffer[ReceiveBufferSize];
     for (auto & s : sockets)
@@ -150,7 +150,7 @@ TEST(TestTcpService, multipleRequestResponse)
     {
         EXPECT_EQ(0, close(s));
     }
-    EXPECT_EQ(sockets.size(), taskExecuteCallsCount.load());
+    EXPECT_EQ(sockets.size(), static_cast<size_t>(taskExecuteCallsCount.load()));
 }
 
 // TODO
